create baseMediaConnection
