+++ src/client/index.ts
@@ -1,5 +1,4 @@
 export * from './packet';
-export * from './stream';
 export * from './voice';
 export * from "./StreamOpts";
 export * from "./GatewayOpCodes";
